Updates to coinbase products and tests after PR comments
