+++ src/assets/js/data/feed/coinbase/formatProducts.js
@@ -1,17 +1,6 @@
-import d3 from 'd3';
 import model from '../../../model/model';
 
-export default function(minute1, minute5, hour1, day1, callback) {
-    d3.json('https://api.exchange.coinbase.com/products', function(error, response) {
-        if (error) {
-            callback(error);
-        } else {
-            callback(error, formatProducts(minute1, minute5, hour1, day1, response));
-        }
-    });
-}
-
-export function formatProducts(minute1, minute5, hour1, day1, response) {
+export default function(minute1, minute5, hour1, day1, response) {
     var products = response.map(function(product) {
         if (product.id !== 'BTC-USD') {
             return {
